### 2026-03-17 #10217 @editorialbot
[PRE REVIEW]: ESPResSo++: A Fast and Extensible Molecular Simulation Package for Coarse-Grained Models
**Submitting author:** @junghans (<a href="http://orcid.org/0000-0003-0925-1458">Christoph Junghans</a>)
**Repository:** https://github.com/espressopp/espressopp
**Branch with paper.md** (empty if default branch):
**Version:** v3.1.0
