### 2026-03-17 #10217 @junghans
> [@junghans](https://github.com/junghans) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

As reviewers I suggest: jngrad, al42and, RudolfWeeber, jglaser and alexxy
