### 2026-08-04 #10217 @diehlpk
Hi @jngrad do you have time to review the paper?
