### 2026-08-12 28eb05d
Adding CONTRIBUTING.md
