### 2026-08-15 #11072 @tingyu66
Hi editor @diehlpk and @junghans, I tested the software using the provided Docker container, and was able to run the provided Python examples.

The submission is for version 3.1, but both the Docker build and CMakeLists.txt report version 3.0.0. Please ensure that the container, source code, and submitted release consistently identify the reviewed version. In general, the paper is well written and contains all the sections required for a JOSS submission. I only have the following minor comments:

- The section titled “New Features since last release” is ambiguous because releases 3.0 and 3.1 already exist. If the comparison is with the previous ESPResSo++ 2.0 publication, please rename it to “Changes since ESPResSo++ 2.0” and state this explicitly.
- Replace “Each C++ class provides a static registerPython() method” with “Each Python-exposed C++ class provides a static registerPython() method.”
- Citation on the footnote might have some formatting issue. It shows "Journal of Open Source Software, ¿VOL?(¿ISSUE?)" on my end.
- MD time calculation in "Example usage" section: The reported md time in the print statement appears incorrect because the first ten iterations use dt = 0.0001, while subsequent iterations use dt = 0.005.
