### 2026-08-15 #11072 @junghans
@tingyu66 thanks for the review, I have hopefully addressed all your comments in https://github.com/espressopp/espressopp/pull/510.

> The submission is for version 3.1, but both the Docker build and CMakeLists.txt report version 3.0.0.

Absolutely, we should have bumped the version in `CMakeLists.txt` to 3.1.0, done now.

> * The section titled “New Features since last release” is ambiguous because releases 3.0 and 3.1 already exist. If the comparison is with the previous ESPResSo++ 2.0 publication, please rename it to “Changes since ESPResSo++ 2.0” and state this explicitly.

You are right, the section title is ambiguous. We are referring to changes since v2.0, as stated in the text later. To make that more clear I updated the section title.

> * Replace “Each C++ class provides a static registerPython() method” with “Each Python-exposed C++ class provides a static registerPython() method.”

Done

> * Citation on the footnote might have some formatting issue. It shows "Journal of Open Source Software, ¿VOL?(¿ISSUE?)" on my end.

I think that will fix itself once the paper gets published.

> * MD time calculation in "Example usage" section: The reported md time in the print statement appears incorrect because the first ten iterations use dt = 0.0001, while subsequent iterations use dt = 0.005.

Thanks for the comment, `dt = 0.0001` is used during warm up and then later `dt = 0.005` is used for the actual simulation. I have updated the example to hopefully make that more clear.
