### 2026-08-16 b06f193
Added a new scientfici application and reference to polymer concepts in Polysomes organization
