### 2026-08-16 b233df3
Fix citation format for Guzman reference
