### 2026-08-16 23aee0e
paper.md: fix some reference
