### 2026-08-26 e6ca14b
CONTRIBUTING.md: fix security email and use https
