### 2026-08-26 ec606bc
installation.rst: fix export PYTHONPATH
