### 2026-08-26 d487dc0
paper.md: fix spaces around cites
