+++ paper/paper.md
@@ -62,7 +62,7 @@ bibliography: paper.bib
 
 # Summary
 
-**ESPResSo++** is an open-source software package for molecular dynamics (MD) simulations with a particular emphasis on coarse-grained (CG) models of soft matter systems[@Praprotnik2008]. Written in C++ with a flexible Python interface, it is designed for high-performance computing (HPC) environments and supports massively parallel simulations through MPI. The package enables simulations of polymers, membranes, colloids and complex fluids with a wide range of interaction models and advanced algorithms.
+**ESPResSo++** is an open-source software package for molecular dynamics (MD) simulations with a particular emphasis on coarse-grained (CG) models of soft matter systems [@Praprotnik2008]. Written in C++ with a flexible Python interface, it is designed for high-performance computing (HPC) environments and supports massively parallel simulations through MPI. The package enables simulations of polymers, membranes, colloids and complex fluids with a wide range of interaction models and advanced algorithms.
 
 ESPResSo++ builds upon the experience of its predecessor [ESPResSo](https://espressomd.org), but provides a cleaner, more modular codebase and enhanced extensibility. It is actively developed by an international community of researchers across multiple institutions and disciplines.
 
@@ -80,7 +80,7 @@ Molecular dynamics simulations are essential tools for exploring the behavior of
 
 Several established MD packages serve the soft matter and coarse-grained simulation communities. **LAMMPS** [@Thompson2022] is a general-purpose, highly scalable code with broad force field support and a flexible plugin architecture; however, its input scripting language is less suited to complex CG workflows, and adaptive resolution is not a core feature. **GROMACS** [@Abraham2015] provides exceptional performance for standard force fields and supports the Martini CG model ecosystem, but adding truly novel CG interactions requires modifying the C++ source, and its file-based workflow is less interactive than a Python API. **HOOMD-blue** [@Anderson2020] is the most directly comparable package, offering a Python-native API and GPU-accelerated CG simulations for soft matter; however, it lacks support for adaptive resolution methods. **OpenMM** [@Eastman2017] excels at GPU-accelerated simulations and allows custom force definitions via algebraic expressions, but targets primarily biomolecular systems and lacks MPI-based multi-node parallelization. The sibling project **ESPResSo** [@Weik2019] shares common roots but focuses on charged systems, hydrodynamic coupling (lattice Boltzmann), and electrokinetics rather than multiscale resolution bridging.
 
-ESPResSo++ occupies a distinct niche as a package purpose-built for coarse-grained and multiscale soft matter simulations. Its key differentiator is first-class support for adaptive resolution simulation (AdResS), which allows seamless coupling of atomistic and coarse-grained representations within a single simulation. ESPResSo++ is currently the only actively maintained MD package that provides AdResS as a core feature; none of the other packages listed above offer this capability. It is worth noting that these capabilities were also contributed to existing general-purpose codes[@Fritsch2012; @Nagarajan2013; @Junghans2010], but got removed again later, hence a dedicated package allows tighter integration of multiscale algorithms with the domain decomposition, neighbor list construction, and force computation pipeline — design choices that would be difficult to retrofit into architectures optimized for all-atom throughput.
[…]
