### 2026-08-26 816b959
installation.rst: fix matplotlib link
