### 2026-08-26 b31bfb0
installation.rst: add note about requirements.txt
