+++ doc/sphinx/installation.rst
@@ -22,6 +22,12 @@ Enter this subdirectory
 
    cd espressopp-|version|
 
+ESPResSo++ requires Python 3.7 or newer. All required Python packages are listed in `requirements.txt`. You can install them via:
+
+.. code-block:: bash
+
+   pip3 install -r requirements.txt`
+
 Create the Makefiles using the cmake command. If you don't have it yet, you have to
 install it first. It is available for all major Linux distributions and also for Mac OS X.
 (ubuntu,debian: "apt-get install cmake" or get it from http://www.cmake.org )
