### 2026-08-26 8d52a70
CI: add test for module
