### 2026-08-27 4e921e0
cmake: fix add_custom_command warning
