+++ testsuite/AdResS/FreeEnergyCompensation/CMakeLists.txt
@@ -1,7 +1,7 @@
 if (NOT ${CMAKE_CURRENT_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_BINARY_DIR})
   add_custom_target(FreeEnergyCompensation_testdata ALL)
   foreach(_file table_fec.tab)
-    add_custom_command(TARGET FreeEnergyCompensation_testdata COMMAND ${CMAKE_COMMAND} -E create_symlink
+    add_custom_command(TARGET FreeEnergyCompensation_testdata POST_BUILD COMMAND ${CMAKE_COMMAND} -E create_symlink
       ${CMAKE_CURRENT_SOURCE_DIR}/${_file} ${CMAKE_CURRENT_BINARY_DIR}/${_file})
   endforeach()
 endif()
