+++ testsuite/FileIOTests/dump_xtc/CMakeLists.txt
@@ -1,7 +1,7 @@
 if (NOT ${CMAKE_CURRENT_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_BINARY_DIR})
   add_custom_target(dump_xtc_testdata ALL)
   foreach(_file expected_with_compression.xtc expected_without_compression.xtc) 
-    add_custom_command(TARGET dump_xtc_testdata COMMAND ${CMAKE_COMMAND} -E create_symlink
+    add_custom_command(TARGET dump_xtc_testdata POST_BUILD COMMAND ${CMAKE_COMMAND} -E create_symlink
       ${CMAKE_CURRENT_SOURCE_DIR}/${_file} ${CMAKE_CURRENT_BINARY_DIR}/${_file})
   endforeach()
 endif()
