+++ testsuite/pi_water/CMakeLists.txt
@@ -2,7 +2,7 @@ if (NOT ${CMAKE_CURRENT_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_BINARY_DIR})
   add_custom_target(pi_water_testdata ALL)
   foreach(_file conf.gro topol.top spcefmatch_tabbond.itp table_O_O.xvg
       table_H_O.xvg table_H_H.xvg table_a1.xvg table_b1.xvg pathintegral.py)
-      add_custom_command(TARGET pi_water_testdata COMMAND ${CMAKE_COMMAND} -E create_symlink
+      add_custom_command(TARGET pi_water_testdata POST_BUILD COMMAND ${CMAKE_COMMAND} -E create_symlink
           ${CMAKE_CURRENT_SOURCE_DIR}/${_file} ${CMAKE_CURRENT_BINARY_DIR}/${_file})
   endforeach()
 endif()
