+++ testsuite/ewald/CMakeLists.txt
@@ -1,7 +1,7 @@
 if (NOT ${CMAKE_CURRENT_SOURCE_DIR} STREQUAL ${CMAKE_CURRENT_BINARY_DIR})
  add_custom_target(ewald_eppDeserno_comparison_testdata ALL)
   foreach(_file deserno_ewald.dat ini_struct_deserno.dat)
-    add_custom_command(TARGET ewald_eppDeserno_comparison_testdata COMMAND ${CMAKE_COMMAND} -E create_symlink
+    add_custom_command(TARGET ewald_eppDeserno_comparison_testdata POST_BUILD COMMAND ${CMAKE_COMMAND} -E create_symlink
       ${CMAKE_CURRENT_SOURCE_DIR}/${_file} ${CMAKE_CURRENT_BINARY_DIR}/${_file})
   endforeach()
 endif()
