### 2026-08-28 #11072 @junghans
@jngrad I believe I took care of all your suggestion above, please have another look if I missed anything. Details are in https://github.com/espressopp/espressopp/issues/514.
