### 2026-08-28 #11072 @jngrad
> [@jngrad](https://github.com/jngrad) I believe I took care of all your suggestions above, please have another look if I missed anything. Details are in [espressopp/espressopp#514](https://github.com/espressopp/espressopp/issues/514).

Thank you for the great work!

More tidbits:
* In `doc/sphinx/index.rst`, consider replacing `instalation.rst` by `installation.rst` (missing letter 'l') to make the installation instructions appear on the website. It is spelled correctly in `doc/sphinx/index_latex.rst`, which is probably why the installation instructions only appeared in the PDF. From the [latest logfile](https://github.com/espressopp/espressopp/actions/runs/33128969152/job/98713754526):
  ```
  /__w/espressopp/espressopp/espressopp/doc/sphinx/index.rst:10: WARNING: toctree contains reference to nonexisting document 'instalation' [toc.not_readable]
  /__w/espressopp/espressopp/espressopp/doc/sphinx/sphinxext/ipython_console_highlighting.py:55: SyntaxWarning: "\[" is an invalid escape sequence.
  Such sequences will not work in the future. Did you mean "\\["? A raw string is also an option.
  input_prompt = re.compile("(In \[[0-9]+\]: )|(   \.\.\.+:)")
  ```
  Sphinx flag `-W` can turn warnings into errors, but some warnings might be unavoidable, such as `index_latex.rst: WARNING: document isn't included in any toctree` in the html output. Maybe `${SPHINX_EXECUTABLE} -b html -w errors_html.log` followed by
  ```sh
  sed -i '/index_latex.rst/d' errors_html.log
  cat errors_html.log
  if [ -s errors_html.log ]; then exit 1; fi
  ```
  could be used to fail the workflow when the list of warnings is not empty?
* In `doc/sphinx/installation.rst`, the line ``pip3 install -r requirements.txt` `` has a trailing backtick that opens a subshell
* Command `git grep espressopppp -- .` shows several namespaces with 2 extra letters 'p'
