### 2026-08-28 a5f1608
installation.rst: fix typo
