### 2026-08-28 0d45755
*.py: delete commented out py2 prints
