+++ doc/sphinx/sphinxext/ipython_console_highlighting.py
@@ -37,7 +37,7 @@ class IPythonConsoleLexer(Lexer):
       In [2]: a
       Out[2]: 'foo'
 
-      In [3]: print a
+      In [3]: print(a)
       foo
 
       In [4]: 1 / 0
