+++ src/System.py
@@ -196,22 +196,17 @@ class SystemLocal(_espressopp.System):
             if len(args) == 1:
                 arg0 = args[0]
                 if isinstance(arg0, Real3D):
-                    #print arg0," is a Real3D object"
                     self.cxxclass.scaleVolume( arg0 )
                 elif hasattr(arg0, '__iter__'):
                     if len(arg0) == 3:
-                        #print args, " has iterator and length 3"
                         self.cxxclass.scaleVolume(self, toReal3DFromVector(arg0) )
                     elif len(arg0) == 1:
-                        #print args, " has iterator and length 1"
                         self.cxxclass.scaleVolume(self, toReal3DFromVector(arg0[0], arg0[0], arg0[0]) )
                     else:
                         print(args, " is invalid")
                 else:
-                    #print args, " is scalar"
                     self.cxxclass.scaleVolume(self, toReal3DFromVector( [arg0, arg0, arg0] ) )
             elif len(args) == 3:
-                #print args, " is 3 numbers"
                 self.cxxclass.scaleVolume(self, toReal3DFromVector(*args) )
             else:
                 print(args, " is invalid")
