### 2026-08-28 #11072 @diehlpk
@junghans `C\texttt{++}` or https://isocpp.org/wiki/faq/misc-environmental-issues#latex-macros makes a nicer looking C++. Not mandatory, but looks nicer.
