### 2026-08-28 #11072 @junghans
> I will let you decide if you want to change the C++, please make a new release and upload the release to Zenodo.

v3.2.0 was released: https://doi.org/10.5281/zenodo.22151265
