### 2026-08-28 #11072 @junghans
> I will let you decide if you want to change the C++, please make a new release and upload the release to Zenodo.

And the ++ is fixed here: https://github.com/espressopp/espressopp/pull/530
