### 2026-08-31 #11072 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
